datastore: use fixed size types for FixedIndexWriter
The file format is independent of the systems pointer size, so it is
better to use fixed size types.

For consistency with most other code, use u64 for the total content
size and u32 for the chunk size. Internally, the latter is also a u64
to avoid some casts and becasue that is what is stored in the header
anyway.

Tested-by: Christian Ebner <c.ebner@proxmox.com>
Link: https://lore.proxmox.com/all/20260210150642.469670-7-r.obkircher@proxmox.com

+++ pbs-datastore/src/datastore.rs
@@ -695,8 +695,8 @@ impl DataStore {
     pub fn create_fixed_writer<P: AsRef<Path>>(
         &self,
         filename: P,
-        size: usize,
-        chunk_size: usize,
+        size: u64,
+        chunk_size: u32,
     ) -> Result<FixedIndexWriter, Error> {
         let full_path = self.inner.chunk_store.relative_path(filename.as_ref());
         FixedIndexWriter::create(full_path, size, chunk_size)
